+++ tests/spec/_noConflictSpec.js
@@ -47,7 +47,7 @@
 
 global.poly2tri = "previous";
 // For the browser, we require the final distributed bundle
-var p2t = (process.browser ? require('../../dist/poly2tri') : require('../../src/poly2tri'));
+var p2t = (process.browser ? require('../../dist/poly2tri.min') : require('../../src/poly2tri'));
 global.poly2tri = p2t;
 
 describe("poly2tri module", function() {