+++ tests/spec/poly2triSpec.js
@@ -41,7 +41,7 @@
 
 
 // For the browser, we require the final distributed bundle
-var p2t = (process.browser ? require('../../dist/poly2tri') : require('../../src/poly2tri'));
+var p2t = (process.browser ? require('../../dist/poly2tri.min') : require('../../src/poly2tri'));
 
 
 /*