+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 [demo]: http://r3mi.github.io/poly2tri.js
 
+- provide a minified build in `dist/poly2tri.min.js`, 
+  using [UglifyJS 2](https://github.com/mishoo/UglifyJS2).
+
 <a name="1.3.1"></a>
 # 1.3.1 (2013-12-24)